refactor(ef): modularize recurrence-rules — extract handlers + event-generator
Splits 659-line recurrence-rules/index.ts into partner-manage-party pattern:
- _handlers/{create,update,pause,resume,cancel}.ts — one file per action
- _lib/types.ts — RecurrenceRule interface, VALID_PATTERNS, Pattern
- _lib/event_generator.ts — generateEvents() + calculateDates()
- index.ts — slim 55-line DISPATCH router (was 659 lines)

Also fixes: TS2352 type cast bugs (rule.parties → unknown intermediate cast)
and adds --allow-read flag to deno.json test task.

Constraint: captureServeHandler test utility imports index.ts directly — router must remain at index.ts
Confidence: high
Scope-risk: narrow

Closes #2147
