+++ supabase/functions/recurrence-rules/_handlers/resume.ts
@@ -1,0 +1,92 @@
+import type { SupabaseClient } from "@supabase/supabase-js";
+import {
+  errorResponse,
+  successResponse,
+} from "../../_shared/response_utils.ts";
+import { requirePartnerPermission } from "../../_shared/partner_permissions.ts";
+import type { RecurrenceRule } from "../_lib/types.ts";
+import { generateEvents } from "../_lib/event_generator.ts";
+
+export async function handleResume(
+  body: Record<string, unknown>,
+  supabase: SupabaseClient,
+  userId: string,
+): Promise<Response> {
+  const ruleId = body.rule_id;
+  if (typeof ruleId !== "string" || !ruleId) {
+    return errorResponse("Missing rule_id", 400);
+  }
+
+  const { data: rule, error: fetchError } = await supabase
+    .from("recurrence_rules")
+    .select(
+      "id, party_id, pattern, days_of_week, month_day, start_time, end_time, end_date, status, last_generated_date, created_at, parties!inner(partner_id)",
+    )
+    .eq("id", ruleId)
+    .maybeSingle();
+
+  if (fetchError) return errorResponse("Failed to load recurrence rule", 500);
+  if (!rule) return errorResponse("Recurrence rule not found", 404);
+
+  if (rule.status !== "paused") {
+    return errorResponse(
+      `Cannot resume a rule with status: ${rule.status}`,
+      400,
+    );
+  }
+
+  const partnerId = (rule.parties as unknown as Record<string, unknown>).partner_id as string;
+  const permCheck = await requirePartnerPermission(supabase, partnerId, userId, ["PARTY_MANAGE"]);
+  if (permCheck) return permCheck;
+
+  // Activate first
+  const { error: updateError } = await supabase
+    .from("recurrence_rules")
+    .update({ status: "active" })
+    .eq("id", ruleId);
+
+  if (updateError) {
+    return errorResponse(`Failed to resume rule: ${updateError.message}`, 500);
+  }
+
+  // Generate missing events: from last_generated_date (or today) → today+30
+  const today = new Date();
+  today.setUTCHours(0, 0, 0, 0);
+
+  let fromDate = today;
+  if (rule.last_generated_date) {
+    const lastGen = new Date(rule.last_generated_date);
+    lastGen.setUTCHours(0, 0, 0, 0);
+    // Generate from day after last_generated_date
+    const nextDay = new Date(lastGen);
+    nextDay.setUTCDate(nextDay.getUTCDate() + 1);
+    if (nextDay < today) fromDate = nextDay;
+  }
+
+  const toDate = new Date(today);
+  toDate.setUTCDate(toDate.getUTCDate() + 30);
+
+  // Build rule object without the joined parties field
+  const ruleData: RecurrenceRule = {
+    id: rule.id,
+    party_id: rule.party_id,
+    pattern: rule.pattern,
+    days_of_week: rule.days_of_week,
+    month_day: rule.month_day,
+    start_time: rule.start_time,
+    end_time: rule.end_time,
+    end_date: rule.end_date,
+    status: "active",
+    last_generated_date: rule.last_generated_date,
+    created_at: rule.created_at,
+  };
+
+  const eventsCreated = await generateEvents(
+    supabase,
+    ruleData,
+    fromDate,
+    toDate,
+  );
+
+  return successResponse({ success: true, events_created: eventsCreated });
+}
