+++ supabase/functions/recurrence-rules/deno.json
@@ -5,6 +5,6 @@
     "@std/testing/mock": "jsr:@std/testing@1/mock"
   },
   "tasks": {
-    "test": "deno test --allow-env --allow-net"
+    "test": "deno test --allow-env --allow-net --allow-read"
   }
 }
